Move permissions-deleting logic out of handler
[#151615607] Refactor CredentialsController to avoid interacting with CredentialDataService directly

+++ src/main/java/io/pivotal/security/service/PermissionService.java
@@ -3,6 +3,8 @@ package io.pivotal.security.service;
 import io.pivotal.security.auth.UserContext;
 import io.pivotal.security.data.PermissionsDataService;
 import io.pivotal.security.entity.CredentialName;
+import io.pivotal.security.exceptions.EntryNotFoundException;
+import io.pivotal.security.exceptions.InvalidAclOperationException;
 import io.pivotal.security.request.PermissionEntry;
 import io.pivotal.security.request.PermissionOperation;
 import org.apache.commons.lang3.StringUtils;
@@ -11,6 +13,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
+import static io.pivotal.security.request.PermissionOperation.WRITE_ACL;
 
 @Service
 public class PermissionService {
@@ -57,7 +61,16 @@ public class PermissionService {
     return permissionsDataService.getAccessControlList(credentialName);
   }
 
-  public boolean deleteAccessControlEntry(String credentialName, String actor) {
+  public boolean deleteAccessControlEntry(UserContext userContext,
+      String credentialName, String actor) {
+    if (!this.hasPermission(userContext.getAclUser(), credentialName, WRITE_ACL)) {
+      throw new EntryNotFoundException("error.credential.invalid_access");
+    }
+
+    if (!this.userAllowedToOperateOnActor(userContext, actor)) {
+      throw new InvalidAclOperationException("error.acl.invalid_update_operation");
+    }
+
     return permissionsDataService.deleteAccessControlEntry(credentialName, actor);
   }
 }
